Started GCD

+++ sumItUpSansVectors.cpp
@@ -9,13 +9,13 @@ const long MODULUS = 1000000007;
 const int N_MAX = 100000;
 const int Q_MAX = 1000000;
 
-void cycle(unsigned long long numArray[], int cycle, int length);
+void cycle(unsigned long numArray[], int cycle, int length);
 
 int main() {
     /* Enter your code here. Read input from STDIN. Print output to STDOUT */   
 	int N, Q, operations[Q_MAX];
-	unsigned long long numbers[N_MAX];
-	unsigned long long output = 0;
+	unsigned long numbers[N_MAX];
+	unsigned long output = 0;
 	
 	// Getting number of Integers
 	cin >> N;
@@ -46,8 +46,8 @@ int main() {
     return 0;
 }
 
-void cycle(unsigned long long numArray[], int cycle, int length){
-	unsigned long long temp[N_MAX];
+void cycle(unsigned long numArray[], int cycle, int length){
+	unsigned long temp[N_MAX];
 	
 	for(int i = 0; i < length; i++) {
 		temp[i] = numArray[i] + numArray[(i-cycle+length)%length];
